quant_modules, quant_utils merged as one file
